Working exercise page

+++ WorkoutDashboard/src/Application/Workout.Dashboard.Web/BusinessOps/IStrRateOperations.cs
@@ -4,7 +4,7 @@ namespace Workout.Dashboard.Web.BusinessOps
 {
     public interface IStrRateOperations
     {
-        double CalculateAverageStrRate(IEnumerable<dynamic> exercises);
+        decimal CalculateAverageStrRate(IEnumerable<dynamic> exercises);
         decimal CalculateStrRate(decimal lift, int repCount);
         dynamic CalculateTopStrRateExecution(IEnumerable<dynamic> executions);
